+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4042,7 +4042,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KXH5Z3AEC555Q62BMJTAGVZ0"),
+  id: z.string().default("01KXJ7644F1XMBHCPRQCWKKAH2"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),