+++ packages/orq-rc/RELEASES.md
@@ -10829,3 +10829,13 @@ Based on:
 - [typescript v4.12.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.28 - packages/orq-rc
+
+## 2026-07-15 06:27:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.29 - packages/orq-rc